requirements.txt
discord.py==1.2.3
requests==2.22.0